requirements.txt
fastapi
mcp[server]
python-jose
uvicorn